Modified Log to use HashMap
Modified the Log class to utilize a Hashmap<String, ArrayList<Object>> where the ArrayList is the data.
The append method no longer is intended to add a "snapshot" of data but rather to add a single value to the contents.

These changes should help by providing a better method to access the logs of particular, named, variables as well as assist in the
transition from row-major storage to column-major storage. This enables us to store the data as CSV files in column major format easily,
so that we can use the Tableau to plot them.

The conversion methods (asArray(), asArrayList()) are empty as they need to be reimplemented to match the change in data type of contents
in a future commit.