Modified Loggable to be compatible with the new "append()" method in Log
The latest append() method has changed it's functionality as detailed in an earlier commit to instead only append a singular element to the Log, as opposed to a whole "snapshot" of data. For this reason, some methods in Loggable had to be changed in order to allow children classes to still provide "snapshots" in the form of an Array. Since the data is stored in a hashmap, children classes will also now have to provide a String array on construction that represents the keys to the hashmap, or alternatively the labels to the values that the child class will log.