ReStructured Logging Utilities to make Logging easier.
This is a proposal to Issue #7. By creating a "Log Object" class, the key and the data can be passed into the snapshot together, preserving the snapshot idea, as well as making the logging of values easier. This is closest to the 2fity approach and also was something somewhat similar to what @CptJJ used in his Pure Pursuit experimenting in another repo. This is still just a proposal, not really sure if this is the best way to do it but it is a way nonetheless to address Issue #7